Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
python-dotenv==1.0.1